Add outbreak.py
In this commit you will find a mostly functional simulator of ordinary differential equations. The code uses the Gillespie algorithm applied to a system of usual differential equations to construct a stochastic realization of the (otherwise deterministic) trajectory of the ODE.

The code is flexible enough so that a user can code in their own system of differential equations; I include an example function that simulates a rather rapid outbreak in a standard SIR model. Tweaking this to include more classes of individual to model COVID is possible. For example, adding a class of asymptomatic but infectious people who have a different social distancing rate than symptomatic people...

Anyway, this is just a toy and it's enough for random people on twitter to play around with different scenarios.